requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
comtypes==1.1.11
idna==3.3
numpy==1.23.0
pandas==1.4.3
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinauto==0.6.8
requests==2.28.0
six==1.16.0
slacker==0.14.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
